Enforce sorted and non-duplicate sample names and indices.
Sorting of the sample names aims to reduce problems with arbitrary
ordering of the matrices when they are combined, which would result in
ambiguous interpretion of various per-cell indices.

Sorting of the indices aims to enable efficient loading by applications
as they don't have to deal with reordering of columns in the matrix;
they just have to decide whether or not to keep the column.