Reorganized the validation code, tests and docs into a v2 category.
This makes space for the upcoming v3 functionality. The general policy is to
avoid lots of version-specific if's; instead, we just make a separate copy of
any code/test/doc that has any change across major versions.

Also removed all doxygen docs for the step-specific validator functions. It's
too burdensome to guarantee stability of the individual validators anyway, so
we'll be treating it as private from now on. As a result, I removed the 
doxygen action, given that there's nothing to really document anymore.